Adding Environment Variables to CirceCi v2
CircleCi retired their AWS permission settings in v2 so I updated the config file to reflect that. It is a new command in the deploy section to configure the access and secret key before making an upload to S3. I also added a short section talking about Environment Variables in the instructions.